Rename shared utils.ts to vis-utils.ts
