+++ apps/storybook/src/ColorBar.stories.tsx
@@ -1,6 +1,6 @@
 import type { ColorBarProps } from '@h5web/lib';
 import { ColorBar, ScaleType } from '@h5web/lib';
-import { COLOR_SCALE_TYPES } from '@h5web/shared/utils';
+import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-utils';
 import type { Meta, StoryFn, StoryObj } from '@storybook/react';
 
 import FillHeight from './decorators/FillHeight';
