+++ apps/storybook/src/DomainWidget.stories.tsx
@@ -1,6 +1,6 @@
 import type { CustomDomain } from '@h5web/lib';
 import { DomainWidget, ScaleType } from '@h5web/lib';
-import { COLOR_SCALE_TYPES } from '@h5web/shared/utils';
+import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-utils';
 import type { Meta, StoryObj } from '@storybook/react';
 import { useState } from 'react';
