+++ apps/storybook/src/HeatmapMesh.stories.tsx
@@ -7,8 +7,8 @@ import {
   VisCanvas,
 } from '@h5web/lib';
 import { assertDefined } from '@h5web/shared/guards';
-import { COLOR_SCALE_TYPES, toTypedNdArray } from '@h5web/shared/utils';
 import { ScaleType } from '@h5web/shared/vis-models';
+import { COLOR_SCALE_TYPES, toTypedNdArray } from '@h5web/shared/vis-utils';
 import type { Meta, StoryObj } from '@storybook/react';
 import { range } from 'lodash';
 import ndarray from 'ndarray';
