+++ apps/storybook/src/HeatmapVis.stories.tsx
@@ -6,7 +6,7 @@ import {
   ScaleType,
 } from '@h5web/lib';
 import { assertDefined } from '@h5web/shared/guards';
-import { COLOR_SCALE_TYPES, toTypedNdArray } from '@h5web/shared/utils';
+import { COLOR_SCALE_TYPES, toTypedNdArray } from '@h5web/shared/vis-utils';
 import type { Meta, StoryObj } from '@storybook/react';
 import ndarray from 'ndarray';
