+++ apps/storybook/src/LineVisDisplay.stories.tsx
@@ -5,7 +5,7 @@ import {
   LineVis,
   mockValues,
 } from '@h5web/lib';
-import { formatTooltipVal } from '@h5web/shared/utils';
+import { formatTooltipVal } from '@h5web/shared/vis-utils';
 import type { Meta, StoryObj } from '@storybook/react';
 
 import LineVisStoriesMeta, { Default } from './LineVis.stories';
