+++ apps/storybook/src/MatrixVis.stories.tsx
@@ -1,6 +1,9 @@
 import { MatrixVis, mockValues } from '@h5web/lib';
 import type { H5WebComplex } from '@h5web/shared/hdf5-models';
-import { createComplexFormatter, toTypedNdArray } from '@h5web/shared/utils';
+import {
+  createComplexFormatter,
+  toTypedNdArray,
+} from '@h5web/shared/vis-utils';
 import type { Meta, StoryObj } from '@storybook/react';
 import { format } from 'd3-format';
