+++ apps/storybook/src/TiledHeatmapMesh/TiledHeatmapMesh.stories.tsx
@@ -5,8 +5,8 @@ import {
   TiledTooltipMesh,
   VisCanvas,
 } from '@h5web/lib';
-import { COLOR_SCALE_TYPES } from '@h5web/shared/utils';
 import { ScaleType } from '@h5web/shared/vis-models';
+import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-utils';
 import type { Meta, StoryObj } from '@storybook/react';
 
 import FillHeight from '../decorators/FillHeight';
