+++ apps/storybook/src/Toolbar.stories.tsx
@@ -11,8 +11,8 @@ import {
   ToggleGroup,
   Toolbar,
 } from '@h5web/lib';
-import { AXIS_SCALE_TYPES } from '@h5web/shared/utils';
 import type { AxisScaleType } from '@h5web/shared/vis-models';
+import { AXIS_SCALE_TYPES } from '@h5web/shared/vis-utils';
 import { useToggle } from '@react-hookz/web';
 import type { Meta, StoryFn, StoryObj } from '@storybook/react';
 import { useState } from 'react';
