+++ apps/storybook/src/TooltipMesh.stories.tsx
@@ -1,5 +1,5 @@
 import { DefaultInteractions, TooltipMesh, VisCanvas } from '@h5web/lib';
-import { formatTooltipVal } from '@h5web/shared/utils';
+import { formatTooltipVal } from '@h5web/shared/vis-utils';
 import type { Meta, StoryObj } from '@storybook/react';
 
 import FillHeight from './decorators/FillHeight';
