+++ packages/app/src/metadata-viewer/utils.ts
@@ -10,7 +10,7 @@ import type {
   Shape,
 } from '@h5web/shared/hdf5-models';
 import { DTypeClass } from '@h5web/shared/hdf5-models';
-import { formatScalarComplex } from '@h5web/shared/utils';
+import { formatScalarComplex } from '@h5web/shared/vis-utils';
 
 export function renderShape(shape: Shape): string {
   if (shape === null) {
