+++ packages/app/src/vis-packs/core/complex/ComplexToolbar.tsx
@@ -9,7 +9,7 @@ import {
   ToggleBtn,
   Toolbar,
 } from '@h5web/lib';
-import { COLOR_SCALE_TYPES } from '@h5web/shared/utils';
+import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-utils';
 import { MdAspectRatio, MdGridOn } from 'react-icons/md';
 
 import type { HeatmapConfig } from '../heatmap/config';
