+++ packages/app/src/vis-packs/core/heatmap/HeatmapToolbar.tsx
@@ -11,7 +11,7 @@ import {
   ToggleBtn,
   Toolbar,
 } from '@h5web/lib';
-import { COLOR_SCALE_TYPES } from '@h5web/shared/utils';
+import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-utils';
 import { MdAspectRatio } from 'react-icons/md';
 
 import type { ExportFormat, ExportURL } from '../../../providers/models';
