+++ packages/app/src/vis-packs/core/scalar/utils.ts
@@ -5,7 +5,7 @@ import type {
   H5WebComplex,
   PrintableType,
 } from '@h5web/shared/hdf5-models';
-import { formatScalarComplex } from '@h5web/shared/utils';
+import { formatScalarComplex } from '@h5web/shared/vis-utils';
 
 import type { ValueFormatter } from '../models';
