+++ packages/lib/src/index.ts
@@ -95,7 +95,7 @@ export { useVisCanvasContext } from './vis/shared/VisCanvasProvider';
 export type { VisCanvasContextValue } from './vis/shared/VisCanvasProvider';
 
 // Utilities
-export { toTypedNdArray } from '@h5web/shared/utils';
+export { toTypedNdArray } from '@h5web/shared/vis-utils';
 
 export {
   getDomain,
@@ -147,7 +147,7 @@ export { default as ErrorBarsGeometry } from './vis/line/errorBarsGeometry';
 export { default as ErrorCapsGeometry } from './vis/line/errorCapsGeometry';
 
 // Constants
-export { COLOR_SCALE_TYPES, AXIS_SCALE_TYPES } from '@h5web/shared/utils';
+export { COLOR_SCALE_TYPES, AXIS_SCALE_TYPES } from '@h5web/shared/vis-utils';
 export { CAMERA_FAR } from './vis/utils';
 export { INTERPOLATORS } from './vis/heatmap/interpolators';
