+++ packages/lib/src/vis/heatmap/HeatmapVis.tsx
@@ -1,7 +1,7 @@
 import { assertDefined } from '@h5web/shared/guards';
-import { formatTooltipVal, getDims } from '@h5web/shared/utils';
 import type { Domain, NumArray } from '@h5web/shared/vis-models';
 import { ScaleType } from '@h5web/shared/vis-models';
+import { formatTooltipVal, getDims } from '@h5web/shared/vis-utils';
 import type { NdArray } from 'ndarray';
 import type { ReactElement, ReactNode } from 'react';
