+++ packages/lib/src/vis/hooks.ts
@@ -1,7 +1,7 @@
 import { createMemo } from '@h5web/shared/createMemo';
-import { getBounds, getValidDomainForScale } from '@h5web/shared/utils';
 import type { AnyNumArray, Domain } from '@h5web/shared/vis-models';
 import { ScaleType } from '@h5web/shared/vis-models';
+import { getBounds, getValidDomainForScale } from '@h5web/shared/vis-utils';
 import { useRerender, useSyncedRef } from '@react-hookz/web';
 import type { Camera } from '@react-three/fiber';
 import { useFrame, useThree } from '@react-three/fiber';
