+++ packages/lib/src/vis/rgb/utils.ts
@@ -1,5 +1,5 @@
-import { getDims, toTypedNdArray } from '@h5web/shared/utils';
 import type { NumArray } from '@h5web/shared/vis-models';
+import { getDims, toTypedNdArray } from '@h5web/shared/vis-utils';
 import type { NdArray } from 'ndarray';
 import ndarray from 'ndarray';
 import { Data3DTexture, FloatType, RedFormat, UnsignedByteType } from 'three';
