+++ packages/lib/src/vis/surface/SurfaceMesh.tsx
@@ -1,4 +1,4 @@
-import { getDims } from '@h5web/shared/utils';
+import { getDims } from '@h5web/shared/vis-utils';
 import { LinearFilter } from 'three';
 
 import HeatmapMaterial from '../heatmap/HeatmapMaterial';
