+++ packages/shared/rollup.utils.js
@@ -12,7 +12,7 @@ const pkg = JSON.parse(fs.readFileSync(path.resolve(dirname, 'package.json')));
  * The object below tells `rollup-plugin-dts` how to map `@h5web/shared/*`
  * imports to their corresponding declaration files.
  *
- * e.g. @h5web/shared/utils => ../shared/dist-ts/utils.d.ts
+ * e.g. @h5web/shared/vis-utils => ../shared/dist-ts/vis-utils.d.ts
  */
 export const aliasEntries = Object.fromEntries(
   Object.entries(pkg.exports)
